Add files via upload
A multinational company decides to carry out a project based on the Internet of Things (IoT) where it will study different parameters in 2 different locations where its headquarters are located with the use of two sensors.

These sensors will measure four parameters, two for each sensor, between July 1 and 14, 2020, and then analyze the results.

The aim is to provide a report so that decisions can be made about the study. This involves querying the collection to obtain different insights, as well as making insertions or updates to the "sensors_IoT" database.